SB-833: `AbstractArtifactManagementService`
