+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-p2-layout-provider/src/main/java/org/carlspring/strongbox/providers/layout/P2LayoutProvider.java
@@ -20,9 +20,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class P2LayoutProvider
-        extends AbstractLayoutProvider<P2ArtifactCoordinates,
-                                              P2RepositoryFeatures,
-                                              P2RepositoryManagementStrategy>
+        extends AbstractLayoutProvider<P2ArtifactCoordinates, P2RepositoryFeatures, P2RepositoryManagementStrategy>
 {
 
     private static final Logger logger = LoggerFactory.getLogger(P2LayoutProvider.class);
